move commands to individual files
