+++ src/lib/terminal/commands/rm.ts
@@ -1,0 +1,23 @@
+import {getNode} from "../filesystem";
+import type {Command, DirNode} from "../types";
+import {err, ok} from "../utils";
+
+
+const command: Command = {
+    desc: "Remove a file",
+    usage: "[-r] <file>",
+    details: ["-r  remove directories recursively"],
+    fn(args, ctx) {
+        const recursive = args.includes("-r") || args.includes("-rf") || args.includes("-fr");
+        const targets = args.filter(a => !a.startsWith("-"));
+        if (!targets[0]) return err("rm: missing operand");
+        const cwd = getNode(ctx.root, ctx.cwd()) as DirNode;
+        const node = cwd.children[targets[0]];
+        if (!node) return err(`rm: cannot remove '${targets[0]}': No such file or directory`);
+        if (node.type === "dir" && !recursive) return err(`rm: cannot remove '${targets[0]}': Is a directory`);
+        delete cwd.children[targets[0]];
+        return ok([]);
+    },
+};
+
+export default command;
