+++ src/lib/terminal/index.ts
@@ -1,2 +1,3 @@
 export * from "./filesystem";
 export * from "./types";
+export * from "./utils";
